backend/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.12.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
oauthlib==3.3.1
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
requests-oauthlib==2.0.0
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1